# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Bio Open Data Dashboard API",
    "description": "API for accessing light pollution data and Taiwan Biodiversity Information (TBIA) dataset",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "paths": {
...
    "/api/charts/light-pollution/distribution": {
      "get": {
        "summary": "Get light pollution distribution by county and month",
        "description": "Retrieve average light pollution data grouped by county and month within a time range",
        "tags": ["Charts", "Light Pollution"],
        "parameters": [
          {
            "name": "start_time",
            "in": "query",
            "required": true,
            "description": "Start time for data retrieval (ISO 8601 format)",
            "schema": {
              "type": "string",
              "format": "date-time",
              "example": "2024-01-01T00:00:00Z"
            }
          },
          {
            "name": "end_time",
            "in": "query",
            "required": true,
            "description": "End time for data retrieval (ISO 8601 format)",
            "schema": {
              "type": "string",
              "format": "date-time",
              "example": "2024-12-31T23:59:59Z"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved light pollution distribution data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/CountyLightData"
                      }
                    },
                    "time_range": {
                      "$ref": "#/components/schemas/TimeRange"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - invalid parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/charts/light-pollution/timeline": {
      "get": {
        "summary": "Get light pollution timeline for a specific county",
        "description": "Retrieve monthly light pollution averages for a specific county and year",
        "tags": ["Charts", "Light Pollution"],
        "parameters": [
          {
            "name": "county",
            "in": "query",
            "required": true,
            "description": "County name (URL encoded for Chinese characters)",
            "schema": {
              "type": "string",
              "example": "新北市"
            }
          },
          {
            "name": "year",
            "in": "query",
            "required": true,
            "description": "Year to retrieve data for",
            "schema": {
              "type": "integer",
              "minimum": 1900,
              "maximum": 2100,
              "example": 2024
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved light pollution timeline data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "county": {
                      "type": "string",
                      "description": "The requested county"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/MonthlyLightData"
                      }
                    },
                    "year": {
                      "type": "integer",
                      "description": "The requested year"
                    },
                    "total_months": {
                      "type": "integer",
                      "description": "Number of months with available data"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - invalid parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/charts/light-pollution/source-ratio": {
      "get": {
        "summary": "Get light pollution by Taiwan regions",
        "description": "Retrieve average light pollution data for Taiwan's four regions (北區/中區/南區/東區) by month",
        "tags": ["Charts", "Light Pollution"],
        "parameters": [
          {
            "name": "month",
            "in": "query",
            "required": true,
            "description": "Month number (1-12)",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 12,
              "example": 6
            }
          },
          {
            "name": "year",
            "in": "query",
            "required": false,
            "description": "Optional year filter",
            "schema": {
              "type": "integer",
              "minimum": 1900,
              "maximum": 2100,
              "example": 2024
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved regional light pollution data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/AreaLightData"
                      }
                    },
                    "month": {
                      "type": "integer",
                      "description": "The requested month"
                    },
                    "year": {
                      "type": "integer",
                      "nullable": true,
                      "description": "The requested year (if provided)"
                    },
                    "total_regions": {
                      "type": "integer",
                      "description": "Number of regions returned"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - invalid parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "TimeRange": {
        "type": "object",
        "properties": {
          "start": {
            "type": "string",
            "format": "date-time",
            "description": "Start time of the query range"
          },
          "end": {
            "type": "string",
            "format": "date-time",
            "description": "End time of the query range"
          }
        },
        "required": ["start", "end"]
      },
...
      "CountyLightData": {
        "type": "object",
        "properties": {
          "county": {
            "type": "string",
            "description": "County name"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MonthlyLightData"
            }
          }
        },
        "required": ["county", "data"],
        "example": {
          "county": "新北市",
          "data": [
            {
              "month": 1,
              "light_pollution_average": 6.70
            },
            {
              "month": 2,
              "light_pollution_average": 6.85
            },
            {
              "month": 3,
              "light_pollution_average": 6.21
            }
          ]
        }
      },
      "MonthlyLightData": {
        "type": "object",
        "properties": {
          "month": {
            "type": "integer",
            "minimum": 1,
            "maximum": 12,
            "description": "Month number (1-12)"
          },
          "light_pollution_average": {
            "type": "number",
            "format": "double",
            "description": "Average light pollution value for the month"
          }
        },
        "required": ["month", "light_pollution_average"],
        "example": {
          "month": 6,
          "light_pollution_average": 3.96
        }
      },
      "AreaLightData": {
        "type": "object",
        "properties": {
          "area": {
            "type": "string",
            "description": "Taiwan region name",
            "enum": ["北區", "中區", "南區", "東區"]
          },
          "light_pollution_average": {
            "type": "number",
            "format": "double",
            "description": "Average light pollution value for the region"
          }
        },
        "required": ["area", "light_pollution_average"],
        "example": {
          "area": "北區",
          "light_pollution_average": 13.65
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message"
          }
        },
        "required": ["error"]
      }
    }
  },
...
}